+++ aiohttp/web.py
@@ -1,5 +1,6 @@
 import asyncio
 import os
+import signal
 import socket
 import stat
 import sys
@@ -318,10 +319,18 @@ class Application(MutableMapping):
         return "<Application 0x{:x}>".format(id(self))
 
 
+class GracefulExit(SystemExit):
+    code = 1
+
+
+def raise_graceful_exit():
+    raise GracefulExit()
+
+
 def run_app(app, *, host=None, port=None, path=None, sock=None,
             shutdown_timeout=60.0, ssl_context=None,
             print=print, backlog=128, access_log_format=None,
-            access_log=access_logger, loop=None):
+            access_log=access_logger, handle_signals=True, loop=None):
     """Run an app locally"""
     user_supplied_loop = loop is not None
     if loop is None:
@@ -414,12 +423,19 @@ def run_app(app, *, host=None, port=None, path=None, sock=None,
         asyncio.gather(*server_creations, loop=loop)
     )
 
-    print("======== Running on {} ========\n"
-          "(Press CTRL+C to quit)".format(', '.join(uris)))
+    if handle_signals:
+        try:
+            loop.add_signal_handler(signal.SIGINT, raise_graceful_exit)
+            loop.add_signal_handler(signal.SIGTERM, raise_graceful_exit)
+        except NotImplementedError:  # pragma: no cover
+            # add_signal_handler is not implemented on Windows
+            pass
 
     try:
+        print("======== Running on {} ========\n"
+              "(Press CTRL+C to quit)".format(', '.join(uris)))
         loop.run_forever()
-    except KeyboardInterrupt:  # pragma: no cover
+    except (GracefulExit, KeyboardInterrupt):  # pragma: no cover
         pass
     finally:
         server_closures = []